changed migrations

+++ database/migrations/2024_08_24_200350_create_pg_allotments_table.php
@@ -27,7 +27,7 @@ class CreatePgAllotmentsTable extends Migration
             $table->string('quota')->nullable();
             $table->string('category')->nullable();
             $table->string('degree')->nullable();
-            $table->integer('seats')->nullable();
+            $table->bigInteger('seats')->nullable();
             $table->bigInteger('fee')->nullable();
             $table->bigInteger('beds')->nullable();
             $table->bigInteger('stipend_1_year')->nullable();
